comments applied
