+++ Analysis/HistMerger.py
@@ -9,6 +9,7 @@ if __name__ == "__main__":
 
 import Common.Utilities as Utilities
 from Analysis.HistHelper import *
+from Analysis.hh_bbtautau import *
 #all_histograms_inputVar[sample_type][channel][QCDRegion][category][key_name]
 # region A = OS_Iso
 # region B = OS_ANTI ISO
@@ -133,26 +134,6 @@ def QCD_Estimation(histograms, all_samples_list, channel, category, uncName, sca
         return ROOT.TH1D()
         #raise RuntimeError("Unable to estimate QCD")
     return hist_data_B
-
-def AddQCDInHistDict(all_histograms, channels, categories, sample_type, uncNameTypes, all_samples_list, scales, onlyCentral):
-    if 'QCD' not in all_histograms.keys():
-            all_histograms['QCD'] = {}
-    for channel in channels:
-        print(f"adding QCD for channel {channel}")
-        for cat in categories:
-            print(f".. and category {cat}")
-            key =( (channel, 'OS_Iso', cat), ('Central', 'Central'))
-            CompareYields(all_histograms, all_samples_list, channel, cat, 'Central', 'Central')
-            print(f".. and uncNameType Central")
-            all_histograms['QCD'][key] = QCD_Estimation(all_histograms, all_samples_list, channel, cat, 'Central', 'Central')
-            for uncNameType in uncNameTypes:
-                print(f".. and uncNameType {uncNameType}")
-                for scale in scales:
-                    print(f" .. and uncScale {scale}")
-                    if onlyCentral and uncNameType!='Central' and scale!='Central' : continue
-                    key =( (channel, 'OS_Iso', cat), (uncNameType, scale))
-                    CompareYields(all_histograms, all_samples_list, channel, cat, uncNameType, scale)
-                    all_histograms['QCD'][key] = QCD_Estimation(all_histograms, all_samples_list, channel, cat, uncNameType, scale)
 
 
 def fillHistDict(inFileRoot, all_histograms, channels, QCDregions, categories, sample_type, uncNameTypes, histNamesDict,signals,onlyCentral=False):
